requirements.txt
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
psycopg2==2.8.2
django-cors-headers==2.5.2
django-rest-swagger==2.2.0